+++ src/launcher/java/org/truffleruby/launcher/options/CommandLineParser.java
@@ -367,16 +367,7 @@ public class CommandLineParser {
                         config.setOption(OptionsCatalog.EXECUTION_ACTION, ExecutionAction.NONE);
                     } else if (extendedOption.startsWith("log=")) {
                         final String levelString = extendedOption.substring("log=".length());
-
-                        final Level level;
-
-                        if (levelString.equals("PERFORMANCE")) {
-                            level = RubyLogger.PERFORMANCE;
-                        } else {
-                            level = Level.parse(levelString.toUpperCase());
-                        }
-
-                        RubyLogger.LOGGER.setLevel(level);
+                        RubyLogger.setLevel(levelString);
                     } else {
                         // Turn extra options into polyglot options and let
                         // org.graalvm.launcher.Launcher.parsePolyglotOption